TensörAkademiAI/Bot/requirements.txt
python-telegram-bot
datetime
